pyproject.toml
[project]
name = "langchain-gpt-41-to-5-upgrade"
version = "0.1.0"
description = "LangChain Azure OpenAI app — GPT 4.1 to GPT 5 migration"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "openai>=1.60.0",
    "python-dotenv>=1.0.0",
    "azure-identity>=1.17.0",
    "azure-ai-evaluation>=1.0.0",
]
